refactor(architecture-diagram): extract icons and svg modules
Move icon resolution, editable SVG emission, text fitting, and editability validation behind owner modules.

Stack-Id: architecture-diagram-m1-3d8ce7
Stack-Position: 3/4
